Add `CustomToml` to `ExtraStyles`.
The least invasive way of supporting custom TOML themes is to add a new
variant to `ExtraStyles`. I implemented `Hash` for `CustomPalette`
because it's not derivable for `Palette`. Finally, indicating whether a
style is nightly must be stored in the palette extensions struct
because I can't hard code it (the old solution for the built in extra styles).